requirements.txt
pyyaml>=6.0
click>=8.1.0
pefile>=2023.2.7
requests>=2.31.0
pathlib>=1.0.1
colorama>=0.4.6
tqdm>=4.66.0
jinja2>=3.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
msal>=1.25.0
aiohttp>=3.9.0
